Move surface to its own directory
