+++ src_c/surface/meson.build
@@ -1,0 +1,51 @@
+simd_blitters_avx2 = static_library(
+    'simd_blitters_avx2',
+    'simd_blitters_avx2.c',
+    dependencies: pg_base_deps,
+    include_directories: '..',
+    c_args: simd_avx2_flags + warnings_error,
+)
+
+simd_blitters_sse2 = static_library(
+    'simd_blitters_sse2',
+    'simd_blitters_sse2.c',
+    dependencies: pg_base_deps,
+    include_directories: '..',
+    c_args: simd_sse2_neon_flags + warnings_error,
+)
+
+simd_surface_fill_avx2 = static_library(
+    'simd_surface_fill_avx2',
+    'simd_surface_fill_avx2.c',
+    dependencies: pg_base_deps,
+    include_directories: '..',
+    c_args: simd_avx2_flags + warnings_error,
+)
+
+simd_surface_fill_sse2 = static_library(
+    'simd_surface_fill_sse2',
+    'simd_surface_fill_sse2.c',
+    dependencies: pg_base_deps,
+    include_directories: '..',
+    c_args: simd_sse2_neon_flags + warnings_error,
+)
+
+surface = py.extension_module(
+    'surface',
+    [
+        'surface.c',
+        'alphablit.c',
+        'surface_fill.c',
+    ],
+    c_args: warnings_error,
+    include_directories: '..',
+    link_with: [
+        simd_blitters_avx2,
+        simd_blitters_sse2,
+        simd_surface_fill_avx2,
+        simd_surface_fill_sse2,
+    ],
+    dependencies: pg_base_deps,
+    install: true,
+    subdir: pg,
+)
