+++ src_c/surface/simd_blitters_avx2.c
