+++ src_c/surface/simd_blitters_sse2.c
