+++ src_c/surface/simd_surface_fill_sse2.c
